"instrument" abstraction
