+++ packages/front/src/appState.ts
@@ -1,5 +1,10 @@
 import type { MidiParser } from './parser/midiParser'
-import type { PlaybackControl } from './player/playTrack'
+import { playMidi, type PlaybackControl } from './player/playTrack'
+import {
+  startProgressUpdates,
+  stopProgressUpdates,
+  updateProgressBar
+} from './ui/progressBar'
 
 type AppState = {
   playing: boolean
@@ -37,7 +42,7 @@ export function getState(): AppState {
 }
 
 export function setSelectedMidi(midi: MidiParser): void {
-  state.currentPlayback?.pause()
+  setPlayback(false)
   state.selectedMidi = midi
   state.currentPlayback = null
 }
@@ -65,4 +70,47 @@ export function setPercussion(enabled: boolean): void {
 }
 export function getPercussion(): boolean {
   return getState().percussion
+}
+
+export function togglePlayback() {
+  const state = getState()
+  const next = !state.playing
+
+  setPlayback(next)
+}
+
+export async function setPlayback(enable: boolean) {
+  const state = getState()
+  state.playing = enable
+
+  if (!enable) {
+    state.currentPlayback?.pause()
+    stopProgressUpdates()
+    return
+  }
+
+  if (state.audioContext.state === 'suspended') {
+    try {
+      await state.audioContext.resume()
+    } catch (error) {
+      console.error('Failed to resume AudioContext:', error)
+      return
+    }
+  }
+
+  if (state.currentPlayback) {
+    await state.currentPlayback?.resume()
+  } else {
+    const playback = playMidi(
+      state.audioContext,
+      state.gainNode,
+      state.analyserNode,
+      state.selectedMidi,
+      state.selectedWaveform
+    )
+    setCurrentPlayback(playback)
+  }
+
+  updateProgressBar()
+  startProgressUpdates()
 }
